# backend/src/modules/account/auth/interfaces/jwt-tokens-response.interface.ts
import { ApiProperty } from '@nestjs/swagger';

export class JwtTokensResponse {
  @ApiProperty({ description: 'access token' })
  accessToken: string;
  @ApiProperty({ description: 'refresh token' })
  refreshToken: string;
}